+++ packages/yarnpkg-shell/sources/index.ts
@@ -1,15 +1,15 @@
 import {PortablePath, npath, ppath, FakeFS, NodeFS}                                  from '@yarnpkg/fslib';
-import {EnvSegment, ArithmeticExpression, ArithmeticPrimary}                         from '@yarnpkg/parsers';
 import {Argument, ArgumentSegment, CommandChain, CommandLine, ShellLine, parseShell} from '@yarnpkg/parsers';
+import {EnvSegment, ArithmeticExpression, ArithmeticPrimary}                         from '@yarnpkg/parsers';
 import {homedir}                                                                     from 'os';
-
 import {PassThrough, Readable, Writable}                                             from 'stream';
 
+import {ShellError}                                                                  from './errors';
 import * as globUtils                                                                from './globUtils';
-import {Handle, ProcessImplementation, ProtectedStream, Stdio, start, Pipe}          from './pipe';
 import {makeBuiltin, makeProcess}                                                    from './pipe';
+import {Handle, ProcessImplementation, ProtectedStream, Stdio, start, Pipe}          from './pipe';
 
-export {globUtils};
+export {globUtils, ShellError};
 
 export type Glob = globUtils.Glob;
 
@@ -89,7 +89,7 @@ function getFileDescriptorStream(fd: number, type: StreamType, state: ShellState
     } break;
 
     default: {
-      throw new Error(`Bad file descriptor: ${fd}`);
+      throw new ShellError(`Bad file descriptor: "${fd}"`);
     }
   }
 
@@ -205,7 +205,7 @@ const BUILTINS = new Map<string, ShellBuiltin>([
           } break;
 
           default: {
-            throw new Error(`Unsupported redirection type: "${type}"`);
+            throw new Error(`Assertion failed: Unsupported redirection type: "${type}"`);
           }
         }
       }
@@ -347,7 +347,7 @@ async function evaluateVariable(segment: ArgumentSegment & {type: `variable`}, o
         } else if (segment.defaultValue) {
           push((await interpolateArguments(segment.defaultValue, opts, state)).join(` `));
         } else {
-          throw new Error(`Unbound argument #${argIndex}`);
+          throw new ShellError(`Unbound argument #${argIndex}`);
         }
       } else {
         if (Object.prototype.hasOwnProperty.call(state.variables, segment.name)) {
@@ -357,7 +357,7 @@ async function evaluateVariable(segment: ArgumentSegment & {type: `variable`}, o
         } else if (segment.defaultValue) {
           push((await interpolateArguments(segment.defaultValue, opts, state)).join(` `));
         } else {
-          throw new Error(`Unbound variable "${segment.name}"`);
+          throw new ShellError(`Unbound variable "${segment.name}"`);
         }
       }
     } break;
@@ -374,6 +374,7 @@ const operators = {
 async function evaluateArithmetic(arithmetic: ArithmeticExpression, opts: ShellOptions, state: ShellState): Promise<number> {
   if (arithmetic.type === `number`) {
     if (!Number.isInteger(arithmetic.value)) {
+      // ZSH allows non-integers, while bash throws at the parser level (unrecoverable)
       throw new Error(`Invalid number: "${arithmetic.value}", only integers are allowed`);
     } else {
       return arithmetic.value;
@@ -481,11 +482,16 @@ async function interpolateArguments(commandArgs: Array<Argument>, opts: ShellOpt
     if (isGlob) {
       const pattern = interpolated.pop();
       if (typeof pattern === `undefined`)
-        throw new Error(`Assertion failed: Expected a glob pattern to have been set.`);
+        throw new Error(`Assertion failed: Expected a glob pattern to have been set`);
 
       const matches = await opts.glob.match(pattern, {cwd: state.cwd, baseFs: opts.baseFs});
-      if (matches.length === 0)
-        throw new Error(`No file matches found: "${pattern}". Note: Glob patterns currently only support files that exist on the filesystem (Help Wanted)`);
+      if (matches.length === 0) {
+        const braceExpansionNotice = globUtils.isBraceExpansion(pattern)
+          ? `. Note: Brace expansion of arbitrary strings isn't currently supported. For more details, please read this issue: https://github.com/yarnpkg/berry/issues/22`
+          : ``;
+
+        throw new ShellError(`No matches found: "${pattern}"${braceExpansionNotice}`);
+      }
 
       for (const match of matches.sort()) {
         pushAndClose(match);
@@ -639,7 +645,7 @@ async function executeCommandChain(node: CommandChain, opts: ShellOptions, state
       });
     } else {
       if (execution === null)
-        throw new Error(`The execution pipeline should have been setup`);
+        throw new Error(`Assertion failed: The execution pipeline should have been setup`);
 
       // Otherwise, depending on the exaxct pipe type, we either pipe stdout
       // only or stdout and stderr
@@ -682,7 +688,20 @@ async function executeCommandLine(node: CommandLine, opts: ShellOptions, state: 
     state.variables[`?`] = String(newCode);
   };
 
-  setCode(await executeCommandChain(node.chain, opts, state));
+  const executeChain = async (chain: CommandChain) => {
+    try {
+      return await executeCommandChain(chain, opts, state);
+    } catch (error) {
+      if (!(error instanceof ShellError))
+        throw error;
+
+      state.stderr.write(`${error.message}\n`);
+
+      return 1;
+    }
+  };
+
+  setCode(await executeChain(node.chain));
 
   // We use a loop because we must make sure that we respect
   // the left associativity of lists, as per the bash spec.
@@ -696,18 +715,18 @@ async function executeCommandLine(node: CommandLine, opts: ShellOptions, state: 
     switch (node.then.type) {
       case `&&`: {
         if (code === 0) {
-          setCode(await executeCommandChain(node.then.line.chain, opts, state));
+          setCode(await executeChain(node.then.line.chain));
         }
       } break;
 
       case `||`: {
         if (code !== 0) {
-          setCode(await executeCommandChain(node.then.line.chain, opts, state));
+          setCode(await executeChain(node.then.line.chain));
         }
       } break;
 
       default: {
-        throw new Error(`Unsupported command type: "${node.then.type}"`);
+        throw new Error(`Assertion failed: Unsupported command type: "${node.then.type}"`);
       } break;
     }
 
@@ -766,7 +785,7 @@ function locateArgsVariableInArgument(arg: Argument): boolean {
     } break;
 
     default:
-      throw new Error(`Unreacheable`);
+      throw new Error(`Assertion failed: Unsupported argument type: "${(arg as Argument).type}"`);
   }
 }